bug fix for handling delete tag and category

+++ app/Http/Controllers/CategoriesController.php
@@ -16,7 +16,7 @@ class CategoriesController extends Controller
      */
     public function index()
     {
-        return view('categories.index')->with('categories',Category::all());
+        return view('categories.index')->with('categories', Category::all());
     }
 
     /**
@@ -39,7 +39,7 @@ class CategoriesController extends Controller
     {
         // $this->validate($request,['name' => 'required|unique:categories']);
         Category::create(['name'=>$request->name]);
-        session()->flash('success','Category created successfully.');
+        session()->flash('success', 'Category created successfully.');
         return redirect(route('categories.index'));
     }
 
@@ -62,7 +62,7 @@ class CategoriesController extends Controller
      */
     public function edit(Category $category)
     {
-        return view('categories.create')->with('category',$category);
+        return view('categories.create')->with('category', $category);
     }
 
     /**
@@ -75,7 +75,7 @@ class CategoriesController extends Controller
     public function update(UpdateCategoriesRequest $request, Category $category)
     {
         $category->update(['name'=>$request->name]);
-        session()->flash('success','Category updated successfully');
+        session()->flash('success', 'Category updated successfully');
         return redirect(route('categories.index'));
     }
 
@@ -87,8 +87,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted because it has some posts');
+            return redirect()->back();
+        }
         $category->delete();
-        session()->flash('success',"Category deleted successfully");
+        session()->flash('success', "Category deleted successfully");
         return redirect(route('categories.index'));
     }
 }
