+++ app/Http/Controllers/TagsController.php
@@ -87,6 +87,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted because it has associated to some posts');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success', "Tag deleted successfully");
         return redirect(route('tags.index'));
